feat(service): refactor stylus colors parsing
