+++ inspect-extension/components/script-json/components/list.mpx
@@ -1,6 +1,10 @@
 <template>
   <view class="list">
-    <view wx:for="{{listData}}" wx:key="*this">{{item}}</view>
+    <view
+      wx:for="{{ listData }}"
+      wx:key="*this"
+      >{{ item }}</view
+    >
   </view>
 </template>
 
@@ -9,17 +13,18 @@ import { createComponent } from '@mpxjs/core'
 
 createComponent({
   data: {
-    listData: ['手机', '电视', '电脑']
-  }
+    listData: ['手机', '电视', '电脑'],
+  },
 })
 </script>
 
 <style lang="stylus">
-  #fff 
-    background #aaa
-  #bbb #ccc
-    border 1px solid #683434
-    background  #333
+#fff
+  background #aaa
+
+#bbb #ccc
+  border 1px solid #683434
+  background #333
 </style>
 
 <script type="application/json">
