+++ inspect-extension/components/script-json/components/list1.mpx
@@ -1,6 +1,10 @@
 <template>
   <view class="list">
-    <view wx:for="{{listData}}" wx:key="*this">{{item}}</view>
+    <view
+      wx:for="{{ listData }}"
+      wx:key="*this"
+      >{{ item }}</view
+    >
   </view>
 </template>
 
@@ -9,8 +13,8 @@ import { createComponent } from '@mpxjs/core'
 
 createComponent({
   data: {
-    listData: ['手机', '电视', '电脑']
-  }
+    listData: ['手机', '电视', '电脑'],
+  },
 })
 </script>
